AI model/requirements.txt
streamlit-authenticator
streamlit
streamlit-chat
openai 
toml
python-decouple
streamlit_extras
langchain
flask
openai
torch
